requirements.txt
# Core application libraries
schedule>=1.2.0
PyYAML>=6.0
Jinja2>=3.1.0
httpx>=0.24.0
markdown>=3.4.0
pandas>=2.0.0

# LLM provider libraries
openai>=1.0.0
google-generativeai>=0.3.0

# PubMed access library
biopython>=1.81

# GUI dependencies
psutil>=5.9.0

# Cross-platform dependencies
platformdirs>=3.0.0

# Windows specific dependencies
pywin32>=305; sys_platform == "win32"

# Security and encryption
cryptography>=41.0.0

# Logging and monitoring
structlog>=23.0.0

# Development dependencies (optional)
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0